Fixed too strict uri-reference character type rule. This commit also contains some (non-functional) marker code (OK, it renders, but crash on updates)

+++ src/svg/svg-types.h
@@ -34,8 +34,8 @@ enum {
 };
 
 struct _SPSVGLength {
-	unsigned long set : 1;
-	unsigned long unit : 4;
+	unsigned int set : 1;
+	unsigned int unit : 4;
 	float value;
 	float computed;
 };
